Refactor : recommendation 기능을 policy 도메인으로 변경
