+++ src/main/java/com/gagechaeum/backend/policy/controller/PolicyController.java
@@ -1,9 +1,9 @@
-package com.gagechaeum.backend.recommendation.controller;
+package com.gagechaeum.backend.policy.controller;
 
 import com.gagechaeum.backend.common.response.CustomResponse;
 import com.gagechaeum.backend.common.response.ResponseCode;
-import com.gagechaeum.backend.recommendation.dto.response.PolicyRecommendationResponseDTO;
-import com.gagechaeum.backend.recommendation.service.RecommendationService;
+import com.gagechaeum.backend.policy.dto.response.PolicyRecommendationResponseDTO;
+import com.gagechaeum.backend.policy.service.PolicyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,17 +11,19 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/recommendations")
+@RequestMapping("/api/policies")
 @RequiredArgsConstructor
-public class RecommendationController {
-    private final RecommendationService recommendationService;
+public class PolicyController {
 
-    @GetMapping("/policies")
+    private final PolicyService policyService;
+
+    // 정책 추천
+    @GetMapping("/recommendation")
     public CustomResponse<PolicyRecommendationResponseDTO> recommendPolicies(
             @RequestParam Long userId // @AuthenticationPrincipal 대신 @RequestParam 사용
     ) {
         // user.getUserId() 대신 파라미터로 받은 userId를 바로 사용
-        PolicyRecommendationResponseDTO response = recommendationService.getRecommendedPolicies(userId);
+        PolicyRecommendationResponseDTO response = policyService.getRecommendedPolicies(userId);
         return CustomResponse.success(ResponseCode.SUCCESS, response);
     }
 }
