+++ src/main/java/com/gagechaeum/backend/policy/mapper/PolicyMapper.java
@@ -1,12 +1,12 @@
-package com.gagechaeum.backend.recommendation.mapper;
+package com.gagechaeum.backend.policy.mapper;
 
-import com.gagechaeum.backend.recommendation.domain.Policy;
+import com.gagechaeum.backend.policy.domain.Policy;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
 @Mapper
-public interface RecommendationMapper {
+public interface PolicyMapper {
     List<Policy> findRecommendedPoliciesByUserId(@Param("userId") Long userId);
 }
