+++ src/main/java/com/gagechaeum/backend/policy/service/PolicyService.java
@@ -1,0 +1,7 @@
+package com.gagechaeum.backend.policy.service;
+
+import com.gagechaeum.backend.policy.dto.response.PolicyRecommendationResponseDTO;
+
+public interface PolicyService {
+    PolicyRecommendationResponseDTO getRecommendedPolicies(Long userId);
+}
