--- src/main/java/com/gagechaeum/backend/recommendation/service/RecommendationService.java
@@ -1,7 +1,0 @@
-package com.gagechaeum.backend.recommendation.service;
-
-import com.gagechaeum.backend.recommendation.dto.response.PolicyRecommendationResponseDTO;
-
-public interface RecommendationService {
-    PolicyRecommendationResponseDTO getRecommendedPolicies(Long userId);
-}
