+++ src/main/resources/com/gagechaeum/backend/policy/mapper/PolicyMapper.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.gagechaeum.backend.recommendation.mapper.RecommendationMapper">
+<mapper namespace="com.gagechaeum.backend.policy.mapper.PolicyMapper">
 
-    <select id="findRecommendedPoliciesByUserId" resultType="com.gagechaeum.backend.recommendation.domain.Policy" parameterType="long">
+    <select id="findRecommendedPoliciesByUserId" resultType="com.gagechaeum.backend.policy.domain.Policy" parameterType="long">
         SELECT
             p.policy_id AS policyId,
             p.policy_name AS policyName,
